separate file for i_bin_packer
